pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "city-swarm-commander"
version = "1.0.0"
description = "Drone fleet logistics command center OpenEnv environment for the Meta PyTorch Hackathon."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Trishit Bodkhe", email = "trishit726@gmail.com" }
]
dependencies = [
    "openenv-core",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
render = ["matplotlib>=3.7.0"]
dev = ["pytest>=7.0", "httpx>=0.24.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["city_swarm_commander*"]
